Added tilemap support
Simple tilemap lookup, deltatime calculation changed a bit

+++ Bradarria/Bradarria/GameLogic.h
@@ -8,7 +8,7 @@ class GameLogic
 public:
   void InitializeSettings();
   
-  float deltaTime{};
+  double deltaTime{};
 
   int windowWidth = 1024;
   int windowHeight = 576;
@@ -20,9 +20,6 @@ public:
   int zoomLevel = 1; // Keeps track of number of zoom level changes (bigger value for broader view)
   int maxZoomLevel = 3;
   int minZoomLevel = 0;
-
-  // --------------- MOVE TO World.cpp once created --------------------------------
-  // 
 
   sf::Vector2i mousePosition{};
